feat:Login,Logout svg added
