+++ src/components/userProfile/GlobalNavigationBar.module.scss
@@ -9,7 +9,8 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    padding-top: 8px 8px;
+    padding: 8px 8px;
+    border-right: 1px solid #F2F5F8;
 }
 
 .iconContainer {
@@ -19,13 +20,13 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    background-color: #444;
-    border-radius: 4px;
+    border-radius: 6px;
 }
 
 .logoIconContainer{
     background-color: #00BD57;
     margin-bottom: 12px;
+    border-radius: 6px;
     position: relative;
 }
 
@@ -36,6 +37,10 @@
     left: 0;
     width: 100%;
     border-bottom: 1px solid #F2F5F8;
+}
+
+.userIconContainer{
+    background-color: #F2F5F8;
 }
 
 .icon {
@@ -50,4 +55,8 @@
 .main {
     flex-grow: 1;
     padding: 16px;
+}
+
+.bottomIconContainer {
+    margin-top: auto;
 }
