+++ src/components/userProfile/GlobalNavigationBar.tsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import styles from './GlobalNavigationBar.module.scss';
+import {ReactComponent as MapuLogo} from '../../assets/MapuLogo.svg';
+import {ReactComponent as Home} from '../../assets/Home.svg';
+import {ReactComponent as Explore} from '../../assets/Explore.svg';
+import {ReactComponent as User} from '../../assets/User.svg';
+import {ReactComponent as Logout } from '../../assets/Logout.svg';
+
+const SideBar = (props: { children: React.ReactNode }) => {
+    return (
+        <div className={styles.container}>
+            <div className={styles.LeftSideBar}>
+                <div className={`${styles.iconContainer} ${styles.logoIconContainer}`}>
+                    <MapuLogo className={styles.icon} />
+                </div>
+                <Link to="/" className={styles.link}>
+                    <div className={styles.iconContainer}>
+                        <Home className={styles.icon} />
+                    </div>
+                </Link>
+                <Link to="/explore" className={styles.link}>
+                    <div className={styles.iconContainer}>
+                        <Explore className={styles.icon} />
+                    </div>
+                </Link>
+                <Link to="/user" className={styles.link}>
+                    <div className={`${styles.iconContainer} ${styles.userIconContainer}`}>
+                        <User className={styles.icon} />
+                    </div>
+                </Link>
+                <div className={`${styles.iconContainer} ${styles.bottomIconContainer}`}>
+                    <Logout className={styles.icon} />
+                </div>
+            </div>
+            <main className={styles.main}>
+                {props.children}
+            </main>
+        </div>
+    );
+};
+
+export default SideBar;
